requirements.txt
vosk
sounddevice
piper-tts
numpy<2.0.0
requests>=2.32.2
google-generativeai
opencv-python
deepface
tf-keras
transformers==4.39.3
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
python-dotenv
tqdm
